Chore: code maintenance (#3)
* Refactor entity classes for improved readability and consistency

* Add AdminRepository and update ApplicationInitConfig for admin user creation

* Resolve User-related classes according to latest changes in database, update user identification to use email

* Update authentication to use email instead of username for user identification

* Change ID types from Long to Integer in entity classes and update UserController methods accordingly

* Refactor repositories and controllers to use Integer for ID types, improve exception handling, and enhance code readability

* Remove unnecessary blank line in AuthenticationService for improved code clarity

+++ src/main/java/com/visita/repositories/AdminRepository.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.visita.entities.AdminEntity;
 
 @Repository
-public interface AdminRepository extends JpaRepository<AdminEntity, Long> {
-
+public interface AdminRepository extends JpaRepository<AdminEntity, Integer> {
 	Optional<AdminEntity> findByUsername(String username);
 }
